requirements.txt
flask
python-dotenv
psycopg2-binary==2.9.9
gunicorn==20.1.0
flask_sqlalchemy